refactor main and map drawer

+++ include/Dimensions.hpp
@@ -7,8 +7,8 @@ public:
   int x;
   int y;
 
-  float dx; // Alternatively width
-  float dy; // Alternatively height
+  int dx; // Alternatively width
+  int dy; // Alternatively height
 
   // Optional parameters
   int z;
@@ -18,7 +18,7 @@ public:
   std::pair<float, float> gridToOpenGL(int x, int y,
                                        const Dimensions &gridDimensions);
 
-  Dimensions(int x, int y, float dx, float dy) : x(x), y(y), dx(dx), dy(dy) {};
+  Dimensions(int x, int y, int dx, int dy) : x(x), y(y), dx(dx), dy(dy) {};
   friend std::ostream &operator<<(std::ostream &os, const Dimensions &d) {
     os << "Dimensions: { x: " << d.x << ", y: " << d.y
        << ", dx (width): " << d.dx << ", dy (height): " << d.dy
